+++ app/services/app-insights/app-insights-api.service.ts
@@ -2,7 +2,7 @@ import { Location } from "@angular/common";
 import { Injectable } from "@angular/core";
 import { Headers, Http, RequestMethod, RequestOptions, RequestOptionsArgs, Response } from "@angular/http";
 import { AccessToken, RetryableHttpCode, ServerError } from "@batch-flask/core";
-import { AccountService } from "app/services/account.service";
+import { AccountService } from "../batch-account.service";
 import { AdalService } from "app/services/adal";
 import { BatchExplorerService } from "app/services/batch-labs.service";
 import { Constants } from "app/utils";
