+++ app/services/azure-batch/core/batch-http.service.ts
@@ -4,8 +4,8 @@ import { Injectable } from "@angular/core";
 import { HttpRequestOptions, HttpService, ServerError } from "@batch-flask/core";
 import { UrlUtils } from "@batch-flask/utils";
 import { AccountResource } from "app/models";
-import { AccountService } from "app/services/account.service";
 import { AdalService } from "app/services/adal";
+import { AccountService } from "app/services/batch-account.service";
 import { BatchExplorerService } from "app/services/batch-labs.service";
 import { Constants } from "common";
 import { Observable, from, throwError } from "rxjs";
