+++ app/services/monitoring/metrics/insights-metrics.service.ts
@@ -3,8 +3,8 @@ import { Response } from "@angular/http";
 import { Observable, of } from "rxjs";
 
 import { Metric, MetricValue, MonitoringMetricList } from "app/models/monitoring";
-import { AccountService } from "app/services/account.service";
 import { ArmHttpService } from "app/services/arm-http.service";
+import { AccountService } from "app/services/batch-account.service";
 import { ThemeService } from "app/services/themes";
 import { flatMap, map, share } from "rxjs/operators";
 import { CoreCountMetrics } from "./core-count-metrics";
